pyproject.toml
[tool.poetry]
name = "hh_parser"
version = "0.1.0"
description = ""
authors = ["Andrey Dubodelov <a.dubodelov@vk.team>"]

[tool.poetry.dependencies]
python = "3.11"
playwright = "^1.28.0"
pytest-playwright = "^0.3.0"
pytest-bdd = "^6.1.1"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
